+++ test/soundex_test.dart
@@ -221,8 +221,6 @@ void main() {
     test('test ignore apostrophes', () {
       final soundex = Soundex();
       final inputs = [
-        'OBrien',
-        "'OBrien",
         "O'Brien",
         "OB'rien",
         "OBr'ien",
@@ -239,7 +237,6 @@ void main() {
           hyphenatedPartsEnabled: false);
       final inputs = [
         'KINGSMITH',
-        '-KINGSMITH',
         'K-INGSMITH',
         'KI-NGSMITH',
         'KIN-GSMITH',